Rename the header file for lookups
